+++ video.h
@@ -1,10 +1,10 @@
-unsigned long pos;
-unsigned long off;
+unsigned int pos;
+unsigned int off;
 
 void display(char * chr);
 	//this will be called from out ASM code
-void PrintNumber(unsigned long bob);
-	//this prints an unsigned long number to the screen in hexadecimal
+void PrintNumber(unsigned int bob);
+	//this prints an unsigned int number to the screen in hexadecimal
 void put(unsigned char);
 //prints a single character to the screen
 void clear_screen();
